+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.14.0] - 23-09-25
+
+### Fixed
+
+- Fixed an issue on Android where the MediaButtonReceiver would crash the app when it did not find a registered MediaBrowserService instance.
+
+### Added
+
+- Added support for side-loaded metadata tracks on iOS.
+
 ## [2.13.0] - 23-09-15
 
 ### Fixed